backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.11.0
sentence-transformers==2.7.0
google-cloud-firestore==2.14.0
firebase-admin==6.4.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.0
gunicorn==21.2.0
aiohttp==3.9.1
